+++ lib/src/traits/fetching.dart
@@ -1,0 +1,7 @@
+import '../../danela.dart';
+
+abstract interface class Fetching<T extends Object> {
+  void init();
+  void dispose();
+  Future<T> fetch(Request request);
+}