+++ lib/src/traits/has_lifecycle.dart
@@ -1,0 +1,4 @@
+abstract interface class HasLifecycle {
+  void init();
+  void dispose();
+}