+++ lib/src/traits/listenable.dart
@@ -1,0 +1,5 @@
+import '../tools/fetched_data.dart';
+
+abstract interface class Listenable<T extends Object> {
+  Stream<FetchedData<T>> get stream;
+}